+++ src/helpers/utils.ts
@@ -1,5 +1,5 @@
 // SITEVERSION must always be line #2 because it is replaced by line number
-export const SITEVERSION = "1.8.1";
+export const SITEVERSION = "1.8.2";
 export const SITENAME = "Cody Burleson";
 
 /**
@@ -23,4 +23,47 @@ export function extractIdFromDocumentPath(){
       path = path.substring(0, path.length - 1); // then remove trailing slash from the path
   }
   return path.substring( path.lastIndexOf('/') + 1 );
+}
+
+/**
+ * Checks the given HTTP fetch response status for OK. 
+ * This helper method is useful when you want to use try/catch with HTTP fetch.
+ * Since any response is considered a successful response, it will not throw 
+ * an error. To throw an error, we need to actually look at the status of
+ * response.ok. Use this helper method in the following way:
+ * 
+ * try {
+ *     const response = await fetch('/assets/data/site-data.json?v=' + SITEVERSION);
+ *     const json = await response.json();
+ *     handleFetchErrors(response);
+ *     // ...
+ * } catch (err) {
+ *     console.log(err);
+ * }
+ * 
+ * This can also be used with a POST type fetch...
+ * 
+ * try {
+ *     const response = await fetch("https://someapi.com/comments", {
+ *         method: "POST",
+ *         body: JSON.stringify(data),
+ *         headers: {
+ *             "Content-Type": "application/json"
+ *         }
+ *     });
+ *     handleFetchErrors(response);
+ *     console.log(await response.json());
+ * } catch(err) {
+ *     console.log(err);
+ * }
+ * 
+ * For more details, see the following YouTube video: 
+ * [GET/POST Requests in StencilJS with the Fetch API](https://www.youtube.com/watch?v=ZtnQ3lD64Uo)
+ * 
+ * @param response 
+ */
+export function handleFetchErrors(response) {
+  if(!response.ok) {
+      throw new Error(response.statusText);
+  }
 }